create readingsutil
